DXP-1706 Merge product select modifier classes to one
